Extracts config default tests into own test file.
